requirements.txt
absl-py==1.3.0
alembic==1.9.1
altair==4.2.2
anyio==3.6.2
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.1.0
arrow==1.2.3
asgiref==3.6.0
astunparse==1.6.3
attrs==21.2.0
Babel==2.11.0
backcall==0.2.0
bayes-optim==0.2.6
bayesian-optimization==1.4.2
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
cached-property==1.5.2
cachetools==5.2.0
catboost==1.1.1
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.1.1
chart-studio==1.1.0
click==8.0.3
cloudpickle==2.2.1
cmarkgfm==0.8.0
colorama==0.4.6
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.4
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.1.4
django-allauth==0.52.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.17.1
duplicity==0.8.21
entrypoints==0.4
et-xmlfile==1.1.0
fasteners==0.14.1
fastjsonschema==2.16.2
filelock==3.6.0
flatbuffers==23.1.4
fonttools==4.29.1
fqdn==1.5.1
fs==2.4.12
fst-pso==1.8.1
future==0.18.2
FuzzyTM==2.0.5
gast==0.4.0
gensim==4.3.0
gitdb==4.0.10
GitPython==3.1.30
glob2==0.7
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
greenlet==2.0.1
grpcio==1.51.1
gunicorn==20.1.0
gyp==0.1
h5py==3.7.0
html5lib==1.1
httplib2==0.20.2
huggingface-hub==0.11.1
hyperopt==0.2.7
idna==3.3
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==4.6.4
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.2.0
json5==0.9.10
jsonpointer==2.3
jsonschema==4.17.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.5.0
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.6.1
jupyter-server==1.23.4
jupyter_client==7.4.8
jupyter_core==5.0.0
jupyter_server_terminals==0.4.3
jupyterlab==3.5.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.17.0
kaleido==0.2.1
kazam==1.4.5
keras==2.11.0
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==15.0.6.1
lightgbm==3.3.5
llvmlite==0.39.1
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
miniful==0.0.6
mistune==2.0.4
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
mysql-connector-python==8.0.32
nbclassic==0.3.7
nbclient==0.6.3
nbconvert==7.2.7
nbformat==5.7.1
nbgrader==0.8.1
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==3.0
nltk==3.8.1
notebook==6.4.8
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipenv==11.9.0
pkginfo==1.8.2
platformdirs==2.5.1
plotly==5.13.0
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==3.12.4
psutil==5.9.4
ptyprocess==0.7.0
py==1.10.0
py-expression-eval==0.3.14
py4j==0.10.9.7
pyaml==21.10.1
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycups==2.0.1
pydeck==0.8.0
pyDOE==0.3.8
pyFUME==0.2.25
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
Pympler==1.0.1
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-json-logger==2.0.4
python3-openid==3.2.0
pythran==0.10.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
rapidfuzz==2.13.7
readme-renderer==34.0
regex==2022.10.31
reportlab==3.6.8
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.1
roman==3.3
rsa==4.9
scikit-learn==1.2.0
scikit-optimize==0.9.0
scipy==1.8.0
seaborn==0.12.2
SecretStorage==3.3.1
semver==2.13.0
Send2Trash==1.8.1b0
sentence-transformers==2.2.2
sentencepiece==0.1.97
shap==0.41.0
simpful==2.9.0
six==1.16.0
sklearn==0.0
slicer==0.0.7
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sobol-seq==0.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.46
sqlparse==0.4.3
ssh-import-id==5.11
sshtunnel==0.4.0
streamlit==1.17.0
style==1.1.0
sympy==1.9
systemd-python==234
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
terminado==0.13.1
testpath==0.5.0
texttable==1.6.4
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1+cpu
torch-tb-profiler==0.4.1
torchaudio==0.13.1+cpu
torchvision==0.14.1+cpu
tornado==6.2
tqdm==4.57.0
traitlets==5.1.1
transformers==4.25.1
trash-cli==0.17.1.14
twine==3.8.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
update==0.0.1
uri-template==1.2.0
urllib3==1.26.5
usb-creator==0.3.7
validators==0.20.0
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
watchdog==2.2.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.2.2
whitenoise==6.3.0
widgetsnbextension==4.0.5
wordcloud==1.8.2.2
wrapt==1.14.1
xdg==5
xgboost==1.7.3
xkit==0.0.0
xlrd==2.0.1
zipp==1.0.0
